Reworked how PyOpenCL module finds its path
Previously, `pkg_resources` was used to find the location of the PyOpenCL module so that the include path could be constructed.
This commit replaces that by simply using the __file__ and os.path utilities

This change fixes an incompatibility with PyOpenCL and pyinstaller/cxfreeze/etc. When frozen, distutils is not able to resolve the PyOpenCL package directly and subsequently raises a DistributionNotFound error. Since we are the PyOpenCL package, we don't need to be using distutils to resolve it. This issue was [reported](https://lists.tiker.net/pipermail/pyopencl/2014-October/001832.html) by someone else in the pyopencl mailing list as well.