some more tweaks to concurrency checker.
Uses WeakSet to store events so we don't have to worry when they get
destroyed.

The check for out of sync events is also more strict: we now store the
event per (buffer, op) as well as per queue and only check events that
are still valid in both and only for the current buffer.