ci+cd.feat(wheel & pypi): improve and simplify the process and support more platforms
This changes the GitHub action to use cibuildwheel tool to build the wheels, was using manylinux docker instead. By moving to the more generic cibuildwheel, the concrete platform-specific steps to build wheels are encapsulated. This introduction allows building wheels for more os a lot easier (support is TODO). As a side effect, pypy wheels are now built.

Second part is to use the GitHub action for publishing to pypi from pypa with simplification to the checking logic.