Make if_positive keep strides of then_ if known. (#523)
* Make if_positive keep strides of then_ if known.

The previous version of `if_positive` always returns a C-ordered array even if `then_` is not C-ordered. This breaks artificial viscosity in Mirgecom when an array context uses a Fortran style data layout. This change makes the output have the same strides as `then_` if the shape of `then_` is known, otherwise it defaults to use the same layout as `criterion`.

* Increment version

* Resolve Flake8 failures

* Use more idiomatic != in shape comparison

Co-authored-by: Andreas Klöckner <inform@tiker.net>